requirements.txt
django
djangorestframework==3.13.1